Agent: Move PluginRegistry from master/ to puppet/
PluginRegistry was erroneously moved to master/ in 8ed50a4f55.
