+++ monkey/infection_monkey/puppet/plugin_registry.py
